passes float dtype to classification_task constructor to pass test

+++ skdata/tests/test_caltech.py
@@ -4,9 +4,9 @@ from skdata import tasks
 
 def test_Caltech101():
     dset = caltech.Caltech101()
-    tasks.assert_img_classification(*dset.img_classification_task())
+    tasks.assert_img_classification(*dset.img_classification_task(dtype='float'))
 
 
 def test_Caltech256():
     dset = caltech.Caltech256()
-    tasks.assert_img_classification(*dset.img_classification_task())
+    tasks.assert_img_classification(*dset.img_classification_task(dtype='float'))
